renaming types -> variant_utils
